Added Map
Initialized map using the current terrain and implemented necessary controller. ( TODO: merge the controller with the HologramMenuController).

Added holographic rays for the mmap in current vfx and added trigger event.

TODO: Create a new shader for the map body which does not effects from light and arrange the transparency of it.